Fixed bug in Ol/UlListItemRegex that was keeping it from including sublists in any given item.
The regex is meant to match, say " * foo" and following lines until either 1) the end of the paragraph, or 2) the next "(\n) * ".  I had broken #2, completely missing that part of the intent of the regex in the original markdown code.

Just keep saying to yourself: I love regular expressions, I love regular expressions....