Added support for navigation to linked solution documents in the preview window.
Overwrites browser navigation behaviour if a link can be resolved to a document
within current solution. This allows linking to (opening) files in the current solution
when a link is clicked in the preview window. By linking to other Markdown files in
the solution a site (wiki) like experience is achieved.

E.g.
[Home](../Readme.md)
[More](Readme.more.md)
[Code](../MyProject.Web/Properties/AssemblyInfo.cs)